requirements.txt
numpy==2.1.1
matplotlib==3.9.2
scikit-learn==1.5.2
click==8.1.7
pandas==2.2.3
pillow==10.4.0
transformers[torch]==4.45.1
plotly
torch
timm
